requirements.txt
blinker            1.9.0
click              8.1.8
contourpy          1.3.1
cycler             0.12.1
dnspython          2.7.0
Flask              3.1.0
Flask-JWT-Extended 4.7.1
fonttools          4.56.0
itsdangerous       2.2.0
Jinja2             3.1.5
joblib             1.4.2

matplotlib         3.10.1
numpy              2.2.3
packaging          24.2
pandas             2.2.3

PyJWT              2.10.1
pymongo            4.11.1
pyparsing          3.2.1

pytz               2025.1
scikit-learn       1.6.1
scipy              1.15.2

utils              1.0.2
Werkzeug           3.1.3

Flask==2.3.2
pymongo==4.4.1
flask-cors==3.0.10